+++ src/components/tec-button/tec-button.model.ts
@@ -1,0 +1,25 @@
+export enum TecButtonSize {
+  tiny = 'tiny',
+  small = 'small',
+  medium = 'medium',
+  large = 'large',
+  giant = 'giant',
+}
+
+export enum TecButtonMode {
+  rounded = 'rounded',
+  square = 'square',
+  radius = 'radius',
+}
+
+export enum TecButtonColor {
+  solid = 'solid',
+  gradient = 'gradient',
+  outline = 'outline',
+  basic = 'basic',
+}
+
+export enum TecButtonIconMode {
+  left = 'left',
+  right = 'right',
+}