+++ src/models/status.model.ts
@@ -1,0 +1,8 @@
+export enum TecStatus {
+  primary = 'primary',
+  secondary = 'secondary',
+  success = 'success',
+  danger = 'danger',
+  warn = 'warn',
+  info = 'info',
+}